requirements.txt
torch==1.2
torchvision==0.2.0
numpy
pillow
tensorboardX
opencv-python
